chapter 8 merge (#3)
* How to set up Flask-JWT-Extended with our app

* Coding the User model and schema

* How to add a register endpoint to the REST API

* How to add a login endpoint to the REST API

* Protect endpoints by requiring a JWT

* JWT claims and authorization

* How to add logout to the REST API

* Request chaining with Insomnia

* Token refreshing with Flask-JWT-Extended

+++ Project_Your_first_REST_API/resources/item.py
@@ -1,7 +1,6 @@
-import uuid
-from flask import request
 from flask.views import MethodView
 from flask_smorest import abort, Blueprint
+from flask_jwt_extended import jwt_required, get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 
 from db import db
@@ -11,14 +10,19 @@ from schemas import ItemSchema, ItemUpdateSchema
 blp = Blueprint("Items", __name__, description="Item operations")
 
 
-@blp.route("/item/<string:item_id>")
+@blp.route("/item/<int:item_id>")
 class Item(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get(self, item_id):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privileges required.")
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -33,10 +37,12 @@ class Item(MethodView):
 
 @blp.route("/item")
 class ItemList(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))  # This is a list of items
     def get(self):
         return ItemModel.query.all()
 
+    @jwt_required(fresh=True)
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
